[Feat] #1 elo알고리즘

+++ Main/src/main/java/com/Main/domain/record/application/dto/RecordDto.java
@@ -16,7 +16,7 @@ public record RecordDto(
             return new RecordDto(0, 0, 0, 0, 0, 0, 0); // 모든 필드를 기본 값으로 설정
         }
         return new RecordDto(
-                record.getScore(),
+                (int) record.getScore(),
                 record.getParticipant(),
                 record.getVictory(),
                 record.getGoal(),
